scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.14.3
google-cloud-bigquery==3.40.1
streamlit==1.32.0
pandas==2.2.0